+++ client/config/toml.go
@@ -8,11 +8,11 @@ import (
 	"github.com/spf13/viper"
 )
 
-const defaultConfigTemplate = `# This is a TOML config file.
+const DefaultClientConfigTemplate = `# This is a TOML config file.
 # For more information, see https://github.com/toml-lang/toml
 
 ###############################################################################
-###                           Client Configuration                            ###
+###                           Client Configuration                          ###
 ###############################################################################
 
 # The network chain ID
@@ -27,17 +27,33 @@ node = "{{ .Node }}"
 broadcast-mode = "{{ .BroadcastMode }}"
 `
 
-// writeConfigToFile parses defaultConfigTemplate, renders config using the template and writes it to
-// configFilePath.
-func writeConfigToFile(configFilePath string, config *ClientConfig) error {
-	var buffer bytes.Buffer
+var configTemplate *template.Template
+
+func init() {
+	var err error
 
 	tmpl := template.New("clientConfigFileTemplate")
-	configTemplate, err := tmpl.Parse(defaultConfigTemplate)
-	if err != nil {
+	if configTemplate, err = tmpl.Parse(DefaultClientConfigTemplate); err != nil {
+		panic(err)
+	}
+}
+
+// setConfigTemplate sets the custom app config template for
+// the application
+func setConfigTemplate(customTemplate string) error {
+	tmpl := template.New("clientConfigFileTemplate")
+	var err error
+	if configTemplate, err = tmpl.Parse(customTemplate); err != nil {
 		return err
 	}
 
+	return nil
+}
+
+// writeConfigFile renders config using the template and writes it to
+// configFilePath.
+func writeConfigFile(configFilePath string, config interface{}) error {
+	var buffer bytes.Buffer
 	if err := configTemplate.Execute(&buffer, config); err != nil {
 		return err
 	}
@@ -46,7 +62,7 @@ func writeConfigToFile(configFilePath string, config *ClientConfig) error {
 }
 
 // getClientConfig reads values from client.toml file and unmarshalls them into ClientConfig
-func getClientConfig(configPath string, v *viper.Viper) (*ClientConfig, error) {
+func getClientConfig(configPath string, v *viper.Viper) (*Config, error) {
 	v.AddConfigPath(configPath)
 	v.SetConfigName("client")
 	v.SetConfigType("toml")