+++ client/tx/factory.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/cosmos/go-bip39"
 	"github.com/spf13/pflag"
+	"github.com/spf13/viper"
 
 	"cosmossdk.io/math"
 
@@ -48,6 +49,19 @@ type Factory struct {
 
 // NewFactoryCLI creates a new Factory.
 func NewFactoryCLI(clientCtx client.Context, flagSet *pflag.FlagSet) (Factory, error) {
+<<<<<<< HEAD
+=======
+	if clientCtx.Viper == nil {
+		clientCtx.Viper = viper.New()
+	}
+
+	if err := clientCtx.Viper.BindPFlags(flagSet); err != nil {
+		return Factory{}, fmt.Errorf("failed to bind flags to viper: %w", err)
+	}
+
+	signModeStr := clientCtx.SignModeStr
+
+>>>>>>> 6601713eb (feat(client): allow overwritting client.toml (#17513))
 	signMode := signing.SignMode_SIGN_MODE_UNSPECIFIED
 	switch clientCtx.SignModeStr {
 	case flags.SignModeDirect:
@@ -65,18 +79,18 @@ func NewFactoryCLI(clientCtx client.Context, flagSet *pflag.FlagSet) (Factory, e
 	var accNum, accSeq uint64
 	if clientCtx.Offline {
 		if flagSet.Changed(flags.FlagAccountNumber) && flagSet.Changed(flags.FlagSequence) {
-			accNum, _ = flagSet.GetUint64(flags.FlagAccountNumber)
-			accSeq, _ = flagSet.GetUint64(flags.FlagSequence)
+			accNum = clientCtx.Viper.GetUint64(flags.FlagAccountNumber)
+			accSeq = clientCtx.Viper.GetUint64(flags.FlagSequence)
 		} else {
 			return Factory{}, errors.New("account-number and sequence must be set in offline mode")
 		}
 	}
 
-	gasAdj, _ := flagSet.GetFloat64(flags.FlagGasAdjustment)
-	memo, _ := flagSet.GetString(flags.FlagNote)
-	timeoutHeight, _ := flagSet.GetUint64(flags.FlagTimeoutHeight)
+	gasAdj := clientCtx.Viper.GetFloat64(flags.FlagGasAdjustment)
+	memo := clientCtx.Viper.GetString(flags.FlagNote)
+	timeoutHeight := clientCtx.Viper.GetUint64(flags.FlagTimeoutHeight)
 
-	gasStr, _ := flagSet.GetString(flags.FlagGas)
+	gasStr := clientCtx.Viper.GetString(flags.FlagGas)
 	gasSetting, _ := flags.ParseGasSetting(gasStr)
 
 	f := Factory{
@@ -98,10 +112,19 @@ func NewFactoryCLI(clientCtx client.Context, flagSet *pflag.FlagSet) (Factory, e
 		feePayer:           clientCtx.FeePayer,
 	}
 
-	feesStr, _ := flagSet.GetString(flags.FlagFees)
+	feesStr := clientCtx.Viper.GetString(flags.FlagFees)
 	f = f.WithFees(feesStr)
 
+<<<<<<< HEAD
 	gasPricesStr, _ := flagSet.GetString(flags.FlagGasPrices)
+=======
+	tipsStr := clientCtx.Viper.GetString(flags.FlagTip)
+	// Add tips to factory. The tipper is necessarily the Msg signer, i.e.
+	// the from address.
+	f = f.WithTips(tipsStr, clientCtx.FromAddress.String())
+
+	gasPricesStr := clientCtx.Viper.GetString(flags.FlagGasPrices)
+>>>>>>> 6601713eb (feat(client): allow overwritting client.toml (#17513))
 	f = f.WithGasPrices(gasPricesStr)
 
 	f = f.WithPreprocessTxHook(clientCtx.PreprocessTxHook)